# Name only secure ciphers in main.go - SWEET32 